requirements.txt
# Core dependencies
qlib>=0.9.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.7.0
lightgbm>=3.3.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas_market_calendars>=4.0.0
tqdm>=4.64.0
joblib>=1.2.0
